requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
pymongo==4.5.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
python-dotenv==1.0.0
flask==2.3.3
plotly==5.16.1
python-dateutil==2.8.2
pytz==2023.3 